+++ app/src/components/ChatOverview/fragments/ChatOverviewItem.tsx
@@ -9,13 +9,13 @@ import {
   Grid,
   styled,
 } from '@mui/material';
+import { useDateFormat } from '@opentalk/common';
 import { isEmpty } from 'lodash';
 import React from 'react';
 
 import { useAppSelector } from '../../../hooks';
 import { ChatProps } from '../../../store/slices/chatSlice';
 import { selectParticipantById } from '../../../store/slices/participantsSlice';
-import { formatTime } from '../../../utils/timeUtils';
 import ParticipantAvatar from '../../ParticipantAvatar';
 
 const ListItemContainer = styled('div')(({ theme }) => ({
@@ -41,7 +41,10 @@ interface IScopedChatItemProps {
 
 const ChatOverviewItem = ({ chat, onClick }: IScopedChatItemProps) => {
   const participant = useAppSelector(selectParticipantById(chat.id));
+  const date = new Date(chat.lastMessage?.timestamp) ?? Date.now;
+  const formattedTime = useDateFormat(date, 'time');
   const getDisplayName = () => (isEmpty(participant) ? chat.id : participant?.displayName);
+
   const renderPrimaryText = () => (
     <Grid container direction={'row'} spacing={1}>
       <Grid item zeroMinWidth xs>
@@ -50,7 +53,7 @@ const ChatOverviewItem = ({ chat, onClick }: IScopedChatItemProps) => {
         </Typography>
       </Grid>
       <Grid item>
-        <TimeTypography variant={'caption'}>{formatTime(chat.lastMessage?.timestamp)}</TimeTypography>
+        <TimeTypography variant={'caption'}>{formattedTime}</TimeTypography>
       </Grid>
     </Grid>
   );