+++ packages/common/jest.config.ts
@@ -8,4 +8,5 @@ module.exports = {
   moduleNameMapper: {
     '@mui/styled-engine': path.resolve(__dirname, '../../app/node_modules/@mui/styled-engine'),
   },
+  setupFilesAfterEnv: ['./setupTests.ts'],
 };