+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Display only a popup instead of the wrong browser page (!830)
 - Fix display of new voting, while viewing results of the previous ones (!846)
 - Uncheck all checkboxes in the voting popover on every new voting (!842)
+- Centralize date and time format (!831)
 
 ### Stability Improvements